+++ src/Core/Responses/Internal/RelatedResourceCollectionResponse.php
@@ -24,13 +24,16 @@ use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use LaravelJsonApi\Core\Json\Hash;
 use LaravelJsonApi\Core\Resources\JsonApiResource;
-use LaravelJsonApi\Core\Responses\Concerns;
+use LaravelJsonApi\Core\Responses\Concerns\HasEncodingParameters;
+use LaravelJsonApi\Core\Responses\Concerns\HasRelationshipMeta;
+use LaravelJsonApi\Core\Responses\Concerns\IsResponsable;
 
 class RelatedResourceCollectionResponse implements Responsable
 {
 
-    use Concerns\HasRelationshipMeta;
-    use Concerns\IsResponsable;
+    use HasEncodingParameters;
+    use HasRelationshipMeta;
+    use IsResponsable;
 
     /**
      * @var JsonApiResource
@@ -72,7 +75,7 @@ class RelatedResourceCollectionResponse implements Responsable
         $document = $encoder
             ->withRequest($request)
             ->withIncludePaths($this->includePaths($request))
-            ->withFieldSets($this->fieldSets($request))
+            ->withFieldSets($this->sparseFieldSets($request))
             ->withResources($this->related)
             ->withJsonApi($this->jsonApi())
             ->withMeta($this->allMeta())