+++ src/Core/Responses/Internal/ResourceResponse.php
@@ -23,12 +23,14 @@ use Illuminate\Contracts\Support\Responsable;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use LaravelJsonApi\Core\Resources\JsonApiResource;
-use LaravelJsonApi\Core\Responses\Concerns;
+use LaravelJsonApi\Core\Responses\Concerns\HasEncodingParameters;
+use LaravelJsonApi\Core\Responses\Concerns\IsResponsable;
 
 class ResourceResponse implements Responsable
 {
 
-    use Concerns\IsResponsable;
+    use HasEncodingParameters;
+    use IsResponsable;
 
     /**
      * @var JsonApiResource|null
@@ -106,7 +108,7 @@ class ResourceResponse implements Responsable
         $document = $encoder
             ->withRequest($request)
             ->withIncludePaths($this->includePaths($request))
-            ->withFieldSets($this->fieldSets($request))
+            ->withFieldSets($this->sparseFieldSets($request))
             ->withResource($this->resource)
             ->withJsonApi($this->jsonApi())
             ->withMeta($this->meta)